backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.9.2
pytest==8.2.2
httpx==0.27.0
pymongo==4.7.2
redis==5.0.6
psycopg[binary]==3.2.9
SQLAlchemy==2.0.31
python-dotenv==1.0.1
alembic==1.13.2
prometheus-client==0.20.0
PyJWT==2.9.0


